[REFACTOR] Reorganize aside components into subdirectory with collapsible wrapper
- Move Discord, TableOfContents (renamed to Toc), and Viewers components to aside/ subdirectory
- Wrap all aside components with Aside.aside_section for collapsible functionality
- Add 'open' attribute to all components for configurable default state
- Update all LiveView imports to use new Aside.* module paths
- Fix duplicate ID issue in TOC by using separate wrapper ID
- Update test file imports to match new module structure
